refactor: fix circular import
