--- models/__init__.py
@@ -1,3 +1,0 @@
-from .templates import *
-from .base import *
-from .preprocessors import *
