+++ modules/preprocessors/__init__.py
@@ -1,0 +1,2 @@
+from modules.preprocessors.base import BasePreprocessor
+from modules.preprocessors.pdf import PDFPreprocessor
